+++ src/main/java/Telecom/SubscriptionService/SubscriptionServiceApplication.java
@@ -4,10 +4,12 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
 import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.amqp.rabbit.annotation.EnableRabbit;
 
 @SpringBootApplication
 @EnableEurekaClient
 @EnableFeignClients(basePackages = "Telecom.SubscriptionService.feign")
+@EnableRabbit
 public class SubscriptionServiceApplication {
 
 	public static void main(String[] args) {